Level 1 Created + HDRP Errors Fixed
The first bare bones level for the game has been finished. Inside the level are a few extra game objects used to tell the dev team the locations of each player. There is also a capsule to reference the size of a person compared to the size of the level. When I first started working on the project I was getting errors because my HDRP packages needed to be updated. I've fixed the issue on my end however I'm not sure if these fixes will carry over to every developer working on this project since it's in the package manager. If anyone has issues related to HDRP, contact Dylan Thomas preferably by phone number.